权限管理完工，剩下index和物料申请页面。

+++ src/main/java/xyz/kakusei/fams/entity/Permission.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Permission {
-    private Integer id;
+    private Byte id;
     private String name;
     private String expression;
 }
